fix bug and delete unused function
