--- jukijuki/plot_value/confusion_matrix_plot.py
@@ -1,74 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-from sklearn.metrics import confusion_matrix
-from mpl_toolkits.axes_grid1 import make_axes_locatable
-from sklearn.utils.multiclass import unique_labels
-
-
-def plot_confusion_matrix(y_true, y_pred, classes,
-                          normalize=False,
-                          title=None,
-                          cmap=plt.cm.Blues):
-    """
-    Refer to: https://scikit-learn.org/stable/auto_examples/model_selection/plot_confusion_matrix.html
-
-    This function prints and plots the confusion matrix.
-    Normalization can be applied by setting `normalize=True`.
-    """
-    if not title:
-        if normalize:
-            title = 'Normalized confusion matrix'
-        else:
-            title = 'Confusion matrix, without normalization'
-
-    # Compute confusion matrix
-    cm = confusion_matrix(y_true, y_pred)
-    # Only use the labels that appear in the data
-    classes = classes[unique_labels(y_true, y_pred)]
-    if normalize:
-        cm = cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
-        print("Normalized confusion matrix")
-    else:
-        print('Confusion matrix, without normalization')
-
-    # print(cm)
-
-    fig, ax = plt.subplots(figsize=(10, 10))
-    im = ax.imshow(cm, interpolation='nearest', cmap=cmap)
-    tick_marks = np.arange(len(classes))
-    plt.xticks(tick_marks, fontsize=25)
-    plt.yticks(tick_marks, fontsize=25)
-    plt.xlabel('Predicted label', fontsize=25)
-    plt.ylabel('True label', fontsize=25)
-    plt.title(title, fontsize=30)
-
-    divider = make_axes_locatable(ax)
-    cax = divider.append_axes('right', size="5%", pad=0.15)
-    cbar = ax.figure.colorbar(im, ax=ax, cax=cax)
-    cbar.ax.tick_params(labelsize=20)
-
-    # We want to show all ticks...
-    ax.set(xticks=np.arange(cm.shape[1]),
-           yticks=np.arange(cm.shape[0]),
-           # ... and label them with the respective list entries
-           xticklabels=classes, yticklabels=classes,
-           #            title=title,
-           ylabel='True label',
-           xlabel='Predicted label')
-
-    # Rotate the tick labels and set their alignment.
-    plt.setp(ax.get_xticklabels(), ha="right",
-             rotation_mode="anchor")
-
-    # Loop over data dimensions and create text annotations.
-    fmt = '.3f' if normalize else 'd'
-    thresh = cm.max() / 2.
-    for i in range(cm.shape[0]):
-        for j in range(cm.shape[1]):
-            ax.text(j, i, format(cm[i, j], fmt),
-                    fontsize=20,
-                    ha="center", va="center",
-                    color="white" if cm[i, j] > thresh else "black")
-    fig.tight_layout()
-    plt.show()
-    return ax
