+++ jukijuki/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.6"
+__version__ = "0.0.7"
