+++ requirements.txt
@@ -1,21 +1,14 @@
 ###### Requirements without Version Specifiers ######
 catboost
 pytorch_tabnet
-japanize_matplotlib
 xfeat
 category_encoders
-texthero
 transformers
 timm
-fasttext
 
 ###### Requirements with Version Specifiers ######
 albumentations==1.1.0
 tensorflow==2.7.0
 tensorflow-hub==0.12.0
 tensorflow-text == 2.7.3
-torch_optimizer==0.1.0
-folium==0.2.1
 opencv-python-headless<4.3
-plotly==5.5.0
-pyyaml==5.4.1
